# docs/swagger.yaml
basePath: /api
definitions:
  categoryapp.CategoryDto:
    properties:
      categoryId:
        type: integer
      createdAt:
        type: string
      name:
        type: string
      parentId:
        type: integer
      sort:
        type: integer
      updatedAt:
        type: string
    type: object
  categoryapp.CategorySpecDto:
    properties:
      category_id:
        type: integer
      required:
        type: boolean
      sort:
        type: integer
      spec_id:
        type: integer
      spec_name:
        type: string
      values:
        items:
          $ref: '#/definitions/categoryapp.SpecValueDto'
        type: array
    type: object
  categoryapp.SpecValueDto:
    properties:
      name:
        type: string
      sort:
        type: integer
      value_id:
        type: integer
    type: object
  handler.BindSpecReq:
    properties:
      categoryId:
        description: 分类ID
        type: integer
      required:
        description: 是否必填
        type: boolean
      sort:
        description: 排序
        type: integer
      specId:
        description: 规格ID
        type: integer
    required:
    - categoryId
    - specId
    type: object
  handler.CancelOrderRequest:
    properties:
      orderId:
        type: integer
      userId:
        type: integer
    required:
    - orderId
    - userId
    type: object
...
  handler.CreateReq:
    properties:
      name:
        type: string
      parentId:
        type: integer
      sort:
        default: 0
        type: integer
    required:
    - name
    type: object
  handler.OrderListRequest:
    properties:
      page:
        minimum: 1
        type: integer
      pageSize:
        maximum: 100
        minimum: 1
        type: integer
      status:
        description: 可选，若不传则查询所有状态的订单
        enum:
        - 0
        - 1
        - 2
        type: integer
      userId:
        type: integer
    required:
    - userId
    type: object
...
  handler.UpdateReq:
    properties:
      name:
        type: string
      parentId:
        type: integer
      sort:
        type: integer
    type: object
...
  handler.createOrderRequest:
    properties:
      consigneeAddress:
        type: string
      consigneeName:
        type: string
      consigneePhone:
        type: string
      items:
        items:
          $ref: '#/definitions/handler.orderItemRequest'
        minItems: 1
        type: array
    required:
    - consigneeAddress
    - consigneeName
    - consigneePhone
    - items
    type: object
...
  handler.orderItemRequest:
    properties:
      productId:
        type: integer
      quantity:
        minimum: 1
        type: integer
      skuCode:
        type: string
    required:
    - productId
    - quantity
    - skuCode
    type: object
...
  orderapp.OrderDTO:
    properties:
      cancelledAt:
        description: 取消时间
        type: string
      consigneeAddress:
        description: 收货人地址
        type: string
      consigneeName:
        description: 收货人姓名
        type: string
      consigneePhone:
        description: 收货人手机号
        type: string
      createdAt:
        description: 创建时间
        type: string
      items:
        description: 订单明细
        items:
          $ref: '#/definitions/orderapp.OrderItemDTO'
        type: array
      logisticsNo:
        description: 物流单号
        type: string
      orderId:
        description: 订单ID
        type: integer
      orderNo:
        description: 订单编号
        type: string
      paidAt:
        description: 支付时间
        type: string
      status:
        description: 订单状态：1-待支付，2-已支付，3-已发货，4-已完成，5-已取消
        type: integer
      statusName:
        description: 订单状态名称
        type: string
      totalAmount:
        description: 单位：分
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      userId:
        description: 用户ID
        type: integer
    type: object
  orderapp.OrderItemDTO:
    properties:
      price:
        description: 单位：分
        type: integer
      productId:
        type: integer
      productName:
        type: string
      quantity:
        type: integer
      skuCode:
        type: string
      subtotal:
        description: 小计，单位：分
        type: integer
    type: object
  orderapp.OrderListResult:
    properties:
      list:
        items:
          $ref: '#/definitions/orderapp.OrderDTO'
        type: array
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: DDD 架构示例服务（用户 / 商品目录 / 订单）。host 为网关端口，经网关按前缀转发到各服务
  title: Demo Service API
  version: "1.0"
paths:
  /category/{id}:
    delete:
      description: 根据分类ID删除分类
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 分类ID格式错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除分类
      tags:
      - 分类
    put:
      consumes:
      - application/json
      description: 根据分类ID更新分类信息，支持部分更新（未传字段不修改）
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分类信息（均为可选字段）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/categoryapp.CategoryDto'
              type: object
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新分类
      tags:
      - 分类
  /category/bindSpec:
    post:
      consumes:
      - application/json
      description: 建立分类与规格维度的多对多绑定；重复绑定返回错误
      parameters:
      - description: 绑定信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.BindSpecReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/categoryapp.CategorySpecDto'
              type: object
      summary: 绑定规格到分类
      tags:
      - 分类
  /category/create:
    post:
      consumes:
      - application/json
      description: 创建一个新的商品分类；同名分类已存在时返回错误
      parameters:
      - description: 分类信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/categoryapp.CategoryDto'
              type: object
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建分类
      tags:
      - 分类
  /category/findAll:
    get:
      description: 返回全部分类列表（不分页）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/categoryapp.CategoryDto'
                  type: array
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询所有分类
      tags:
      - 分类
  /category/findSpecs/{id}:
    get:
      description: 返回分类下全部绑定的规格维度（含规格值、排序、是否必选）
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/categoryapp.CategorySpecDto'
                  type: array
              type: object
      summary: 查询分类绑定的规格
      tags:
      - 分类
  /category/unbindSpec:
    delete:
      description: 删除分类与规格维度的绑定关系
      parameters:
      - description: 分类ID
        in: query
        name: categoryId
        required: true
        type: integer
      - description: 规格ID
        in: query
        name: specId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 解除规格绑定
      tags:
      - 分类
  /order/cancel:
    post:
      consumes:
      - application/json
      description: 取消订单并回补库存，仅待支付订单可取消，校验订单归属。需要登录（JWT）
      parameters:
      - description: 订单ID与用户ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CancelOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 订单取消成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 订单已取消或当前状态不允许取消
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未登录或 token 失效
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 订单不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 取消订单
      tags:
      - 订单
  /order/create:
    post:
      consumes:
      - application/json
      description: 下单并扣减库存（事务保证），用户ID取自 JWT；同一用户 3 秒内重复提交返回 429。
      parameters:
      - description: 订单信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/orderapp.OrderDTO'
              type: object
        "400":
          description: 参数校验失败/商品已下架/库存不足/收货人信息不完整
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未登录或 token 失效
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 商品或 SKU 不存在
          schema:
            $ref: '#/definitions/response.Response'
        "429":
          description: 请求过于频繁（3 秒防重窗口）
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建订单
      tags:
      - 订单
  /order/list:
    post:
      consumes:
      - application/json
      description: 按用户分页查询订单，可按状态过滤：不传或 0 查全部，1-待支付，2-已支付。需要登录（JWT）
      parameters:
      - description: 查询条件
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.OrderListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/orderapp.OrderListResult'
              type: object
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未登录或 token 失效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 分页查询订单列表
      tags:
      - 订单
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"